Indent fixes, added electric-newline.
We need to do our case detection last(ish) because it is basically a special
kind of block and we want other indent rules to have a chance to match first.

Fn electric-newline does "smart stuff" when you press ENTER. Currently that
includes one smart thing which is adding "* " and indenting properly when you
press newline in the middle of a scaladoc comment.