Fixed issues with single-line blocks.
This new system is getting annoyingly complicated as well. Sigh. Certainly part
of the blame here is Scala and its extra crazy grammar, but all this special
case state pushing and popping is making me feel like there must be a better
way... Maybe the indenter needs to match on multiple states on the top of the
stack... blah.